# swagger.yaml
openapi: 3.0.0
info:
  title: Event Booking API
  version: 1.0.0

paths:
  /events:
    get:
      summary: Get all events
      responses:
        '200':
          description: List of events

    post:
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - date
                - total_capacity
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                total_capacity:
                  type: integer
      responses:
        '200':
          description: Event created

  /bookings:
    post:
      summary: Book tickets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - event_id
                - tickets
              properties:
                user_id:
                  type: integer
                event_id:
                  type: integer
                tickets:
                  type: integer
      responses:
        '200':
          description: Booking successful

  /users/{id}/bookings:
    get:
      summary: Get bookings of a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User bookings list

  /events/{id}/attendance:
    post:
      summary: Mark attendance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - booking_code
              properties:
                booking_code:
                  type: string
      responses:
        '200':
          description: Attendance marked